Refactor: rename ConsoleTextEntry to ConsoleInput and extract SuggestionPopup/CommandPanel
