--- game/addons/menu/Code/DevUI/Console/ConsoleTextEntry.cs.scss
@@ -1,282 +1,0 @@
-@import '/styles/_Theme.scss';
-
-.console-textentry
-{
-	position: relative;
-	z-index: 100;
-	background-color: rgba( $default-800, 0.85 );
-	border: 1px solid transparent;
-	border-radius: $rounding-small;
-	cursor: text;
-	flex-shrink: 0;
-	color: $default-text;
-	overflow: visible;
-	flex-grow: 1;
-	flex-basis: 0;
-	min-width: 180px;
-	min-height: 30px;
-	height: 30px;
-	max-height: 30px;
-	padding: 0 10px;
-	align-items: center;
-	justify-content: flex-start;
-
-	.prefix-label,
-	.suffix-label
-	{
-		display: none;
-	}
-
-	.content-label
-	{
-		position: relative;
-		flex-grow: 1;
-		flex-shrink: 1;
-		min-width: 0;
-		color: $default-text;
-		cursor: text;
-	}
-
-	&:active, &:focus, &:hover
-	{
-		border-color: rgba( $primary-blue, 0.45 );
-	}
-}
-
-.command-panel
-{
-	position: absolute;
-	left: 0;
-	right: 0;
-	top: calc( -52px - 12px );
-	height: 52px;
-	min-height: 52px;
-	max-height: 52px;
-	padding: 8px 14px;
-	border-radius: $rounding-small;
-	border: 1px solid rgba( $default-border, 0.75 );
-	background-color: rgba( $default-900, 1 );
-	z-index: 10001;
-	gap: 0;
-	cursor: default;
-	flex-direction: row;
-	align-items: center;
-	justify-content: center;
-	opacity: 1;
-	transform: translateY( 0px );
-	transition: opacity 0.14s ease, transform 0.14s ease;
-
-	> .header,
-	> .header > .command
-	{
-		cursor: default;
-	}
-
-	> .header
-	{
-		width: calc( 100% - 304px );
-		min-width: 0;
-		flex-direction: row;
-		align-items: center;
-		justify-content: flex-start;
-		height: 100%;
-		margin-right: auto;
-
-		> .command
-		{
-			font-family: Roboto Mono;
-			font-size: 15px;
-			font-weight: 800;
-			color: #ffffff;
-			white-space: nowrap;
-			overflow: hidden;
-			text-overflow: ellipsis;
-			text-align: left;
-			max-width: 100%;
-		}
-	}
-
-	> .controls
-	{
-		position: absolute;
-		right: 14px;
-		top: 50%;
-		transform: translateY( -50% );
-		align-items: center;
-		justify-content: flex-end;
-		gap: 6px;
-		min-height: 24px;
-		max-width: 276px;
-
-		> .arg-btn
-		{
-			min-width: 48px;
-			height: 24px;
-			padding: 0 9px;
-			border-radius: $rounding-small;
-			border: 1px solid rgba( $default-border, 0.6 );
-			background-color: rgba( $default-700, 0.35 );
-			color: #ffffff;
-			font-family: Roboto Mono;
-			font-size: 11px;
-			font-weight: 700;
-			cursor: pointer;
-			transition: background-color 0.08s ease-out, border-color 0.08s ease-out, color 0.08s ease-out;
-
-			&:hover
-			{
-				background-color: rgba( $default-600, 0.42 );
-			}
-
-			&:active
-			{
-				background-color: rgba( $primary-blue, 0.34 );
-				border-color: rgba( $primary-blue, 0.72 );
-				color: #ffffff;
-			}
-
-			&.secondary
-			{
-				color: rgba( $default-200, 0.92 );
-				background-color: rgba( $default-600, 0.26 );
-				min-width: 74px;
-
-				&:active
-				{
-					background-color: rgba( $primary-blue, 0.3 );
-					border-color: rgba( $primary-blue, 0.68 );
-					color: #ffffff;
-				}
-			}
-
-			&.active
-			{
-				background-color: rgba( $primary-blue, 0.3 );
-				border-color: rgba( $primary-blue, 0.64 );
-			}
-		}
-	}
-
-	&:not(.has-actions)
-	{
-		> .controls
-		{
-			display: none;
-		}
-	}
-
-	&.hidden
-	{
-		opacity: 0;
-		transform: translateY( 4px );
-		pointer-events: none;
-	}
-
-	&.is-bool
-	{
-		> .controls > .arg-btn
-		{
-			min-width: 56px;
-		}
-	}
-}
-
-.console-suggestion-menu
-{
-	position: absolute;
-	left: 0;
-	right: 0;
-	top: calc( -240px - 12px );
-	height: 240px;
-	max-height: 240px;
-	min-height: 240px;
-	flex-direction: column;
-	overflow-y: scroll;
-	overflow-x: hidden;
-	gap: 4px;
-	border-radius: $rounding-default;
-	background-color: rgba( $default-900, 1 );
-	border: 1px solid rgba( $default-border, 0.85 );
-	padding: 6px;
-	z-index: 10000;
-	opacity: 1;
-	transform: translateY( 0px );
-	transition: opacity 0.14s ease, transform 0.14s ease;
-
-	&.hidden
-	{
-		opacity: 0;
-		transform: translateY( 6px );
-		pointer-events: none;
-	}
-
-	> .suggestion-item
-	{
-		height: 34px;
-		min-height: 34px;
-		max-height: 34px;
-		cursor: pointer;
-		flex-direction: row;
-		align-items: center;
-		gap: 10px;
-		justify-content: flex-start;
-		padding: 6px 8px;
-		font-family: Roboto Mono;
-		font-weight: 700;
-		border-radius: $rounding-small;
-		background-color: rgba( $default-700, 0.2 );
-		border: 1px solid transparent;
-		transition: background-color 0.06s ease-out;
-		color: #ffffff;
-		overflow: hidden;
-
-		> label
-		{
-			font-size: 12px;
-			line-height: 16px;
-			padding: 0;
-		}
-
-		> .name
-		{
-			flex-grow: 1;
-			flex-shrink: 1;
-			min-width: 0;
-			white-space: nowrap;
-		}
-
-		> .value
-		{
-			flex-grow: 0;
-			flex-shrink: 0;
-			white-space: nowrap;
-			color: rgba( $default-200, 0.8 );
-			text-align: right;
-		}
-
-		&:hover
-		{
-			background-color: rgba( $default-600, 0.24 );
-		}
-
-		&.active
-		{
-			background-color: rgba( $default-500, 0.2 );
-			border-color: rgba( $default-border, 0.9 );
-		}
-
-		&.typed-match
-		{
-			background-color: rgba( $primary-blue, 0.22 );
-			border-color: rgba( $primary-blue, 0.62 );
-
-			> .name
-			{
-			}
-			> .value
-			{
-				color: rgba( $default-200, 0.86 );
-			}
-		}
-	}
-}
